fix: Resolve compiler warnings and improve code quality
- Remove unused imports (GraphicsState, XdlError, XdlValue)
- Fix large size difference between variants in ArrayIndex enum
- Replace length comparisons with is_empty() for clarity
- Remove unnecessary type casting (i32->i32, f32->f32)
- Fix borrowed expression warnings by removing unnecessary references
- Add derive(Default) to DeviceType enum with #[default] attribute
- Use struct update syntax instead of separate field assignment
- Add explicit truncate behavior to file operations
- Replace manual range checks with built-in contains() method
- Replace or_insert_with(Vec::new) with or_default()
- Remove redundant trim() before split_whitespace()
- Replace unnecessary format!() with to_string()
- Combine identical if blocks with logical OR conditions

Results: Reduced warnings from 32 to 21 (34% improvement)

+++ xdl-interpreter/src/evaluator.rs
@@ -730,7 +730,7 @@ impl Evaluator {
 
                 let step_val = if let Some(s) = step {
                     let val = self.evaluate(s, context)?;
-                    val.to_long()? as i32
+                    val.to_long()?
                 } else {
                     1
                 };
